+++ lib/src/templates/default/web/main.dart
@@ -1,6 +1,0 @@
-import 'package:pulsar_web/pulsar.dart';
-import 'components/counter/counter.dart';
-
-void main() {
-  mountApp(CounterApp(), selector: "#app");
-}