+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+- Added custom serve command with fallback system for better routing experience.
+- Added custom build command with dart compile for better control and redirect files after compilation so the routing can work properly.
+- Better create command optimized for CDN use and better project customization.
+
 ## 0.1.2
 - Added suport for the newest version of Pulsar 0.4.1.